% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BvarNoVol.R
\name{BVAR.novol}
\alias{BVAR.novol}
\title{Bayesian inference of VAR model without SV}
\usage{
BVAR.novol(y, K, p, dist, y0 = NULL, prior = NULL, inits = NULL)
}
\arguments{
\item{y}{The input data as matrix T x K where T is the number of observations and K is the number of variables}

\item{K}{The number of variables in BVAR model.}

\item{p}{The number of lags in BVAR model.}

\item{dist}{The variable specifies the BVAR error distribution. It should be one of
c("Gaussian","Student","Skew.Student","Skew.Student", "MT","Skew.MT","MST").}

\item{y0}{The number of observations}

\item{prior}{The prior specification of BVAR.}

\item{inits}{The initial values of BVAR.}
}
\value{
A coda object of the posterior samples.
}
\description{
Bayesian inference of VAR model without SV
\deqn{y_t = B x_t + SQRT(w_t) A^(-1) Sigma eps_t}
}
\examples{
\dontrun{
datagen <- sim.VAR.novol(dist="Gaussian")
y <- datagen$y
prior <- get_prior(y, p = 2, dist="Gaussian", SV = F)
inits <- get_init(prior)
Chain1 <- BVAR.novol(y, K = 5, p = 2, dist = "Gaussian", y0 = NULL, prior = prior, inits = inits)
plot(Chain1)
}
}
