% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ChibLL.R
\name{ChibLLP}
\alias{ChibLLP}
\title{Marginal log likelihood of BVAR model}
\usage{
ChibLLP(Chain, ndraws = NULL, numCores = NULL)
}
\arguments{
\item{Chain}{The fatBVARSV object from command BVAR.}
}
\value{
The marginal log likelihood density of of BVAR model
}
\description{
This function returns a marginal log likelihood density of BVAR-SV-fatTail model.
}
\examples{
\dontrun{
marginal_dens1 <- marginalLL(Chain1)
}

}
