% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fatBVARS-package.R
\docType{package}
\name{fatBVARS-package}
\alias{fatBVARS-package}
\alias{fatBVARS}
\title{Bayesian VAR with Stochastic volatility and fat tails}
\description{
Bayesian VAR with Stochastic volatility and fat tails
}
\details{
\tabular{ll}{
Package: \tab fatBVARS\cr
Type: \tab Package\cr
Version: \tab 0.1.0 \cr
Date: \tab 2019-12-07\cr
License: \tab GPL (>=2)\cr
Depends: \tab R (\eqn{\geq 3.5.0}{>= 3.5.0})\cr
Imports: \tab Rcpp \cr
Suggests: \tab  \cr
LazyLoad: \tab yes }
}
\author{
Hoang Nguyen
}
