% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast.R
\name{forecast_density}
\alias{forecast_density}
\title{Forecast density of future observations of BVAR model}
\usage{
forecast_density(Chain, y_current = NULL, y_obs_future,
  t_current = NULL)
}
\arguments{
\item{Chain}{The fatBVARSV object from command BVAR.}

\item{y_current}{The current values of y.}

\item{y_obs_future}{The future observable values of y.}

\item{t_pred}{The time prediction horizon.}
}
\value{
The forecast density of future observations of BVAR model
}
\description{
This function returns a forecast density of future observations of BVAR-SV-fatTail model.
}
\examples{
\dontrun{
forecast_dens1 <- forecast_density(Chain1, y_obs_future)
}

}
