% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast.R
\name{get_forecast}
\alias{get_forecast}
\title{Forecast future observations of BVAR model}
\usage{
get_forecast(Chain, y0 = NULL, t_pred = 12, t_current = NULL,
  Nfsample = NULL, thin = 1)
}
\arguments{
\item{Chain}{The fatBVARSV object from command BVAR.}

\item{t_pred}{The time prediction horizon.}
}
\value{
The list of forecast future observations of BVAR model
}
\description{
This function returns a forecast future observations of BVAR-SV-fatTail model.
}
\examples{
\dontrun{
forecast1 <- get_forecast(Chain1)
}

}
