% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prior.R
\name{get_init}
\alias{get_init}
\title{Initial values from the prior of BVAR model}
\usage{
get_init(prior, samples = 1100, burnin = 100, thin = 1)
}
\arguments{
\item{prior}{The prior distribution.}
}
\value{
A list of initial values.
}
\description{
This function returns initial values of BVAR-SV-fatTail model from its prior.
}
\examples{
\dontrun{
inits <- get_init(prior)
}

}
