% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impluseResponse.R
\name{get_irf}
\alias{get_irf}
\title{Impulse response function of BVAR model}
\usage{
get_irf(
  Chain,
  impulse.variable = 1,
  response.variable = 2,
  atT = NULL,
  n.ahead = 20,
  draw.plot = FALSE
)
}
\arguments{
\item{Chain}{The fatBVARSV object from command BVAR.}

\item{impulse.variable}{The position of impulse variable}

\item{response.variable}{The position of reponse variable}

\item{atT}{The time at which impulse response function is calculated}

\item{n.ahead}{Maximal time between impulse and response (defaults to 20).}

\item{draw.plot}{Show plot}
}
\value{
A list of two elements: Contemporaneous impulse responses and Matrix of simulated impulse responses
}
\description{
This function returns a impulse response function of BVAR-SV-fatTail model.
}
\examples{
\dontrun{
irf_obj <- get_irf(Chain = Chain, impulse.variable = 1, response.variable = 2,
                   t = NULL, n.ahead = 20, draw.plot = FALSE)
}
}
\references{
bvars package https://cran.r-project.org/web/packages/bvarsv
}
