% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ultis.R
\name{get_post}
\alias{get_post}
\title{Get Posterior samples of BVAR model}
\usage{
get_post(obj, element = NULL, ...)
}
\arguments{
\item{obj}{The Chain/mcmc obtained by fatBVARSVobj$mcmc}

\item{element}{The name of parameters.}
}
\value{
The Posterior samples of BVAR-SV-fatTail model in mcmc object
}
\description{
This function returns Posterior samples of BVAR-SV-fatTail model.
}
\examples{
\dontrun{
B_samples <- get_post(Chain, element = "B")
}
}
