% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prior.R
\name{get_prior}
\alias{get_prior}
\title{Prior distribution of BVAR model}
\usage{
get_prior(
  y,
  p,
  priorStyle = c("Minnesota"),
  dist = c("Gaussian"),
  SV = FALSE,
  ...
)
}
\arguments{
\item{y}{The input data as matrix T x K where T is the number of observations and K is the number of variables}

\item{p}{The number of lags in BVAR model.}

\item{priorStyle}{The prior style in BVAR model should be c("Minnesota", "OLS")}

\item{dist}{The variable specifies the BVAR error distribution. It should be one of
c("Gaussian","Student", "Hyper.Student", "multiStudent","Hyper.multiStudent","orthoStudent").}

\item{SV}{The indicator if this BVAR model has a Stochastic volatility part.}
}
\value{
A list of prior specifications. \eqn{b \sim N(b0, V_b_prior)}, \eqn{a \sim N(0, 1000I)}, \eqn{sigmaSq \sim IG(0.5*sigma_T0, 0.5*sigma_S0)},
\eqn{nu \sim G(a,b)}, \eqn{gamma \sim N(0, I)}.
}
\description{
This function returns a prior specification of BVAR-SV-fatTail model.
}
\examples{
\dontrun{
prior <- get_prior(y, p = 2, dist = c("Student"), SV = FALSE)
}
}
