% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marginalLLSing.R
\name{marginalLLSing}
\alias{marginalLLSing}
\title{Marginal log likelihood of BVAR model}
\usage{
marginalLLSing(Chain, ndraws = NULL)
}
\arguments{
\item{Chain}{The fatBVARSV object from command BVAR.}
}
\value{
The marginal log likelihood density of of BVAR model
}
\description{
This function returns a marginal log likelihood density of BVAR-SV-fatTail model.
}
\examples{
\dontrun{
marginal_dens1 <- marginalLLSing(Chain1)
}

}
