% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ultis.R
\name{plot.fatBVARSV}
\alias{plot.fatBVARSV}
\title{#' @export
lh_post <- function(ytilde, hc, alph, HinvSH_h, t_max, K){
  llike = - 0.5*sum(hc) -
            0.5*t(ytilde) %*% sparseMatrix(i = 1:(t_max*K), j = 1:(t_max*K), x = exp(-hc)) %*% ytilde
  lpost = as.numeric(llike +  -.5*Matrix::t(hc-alph)%*%HinvSH_h%*%(hc-alph))
  return(lpost)
}}
\usage{
\method{plot}{fatBVARSV}(fatBVARSVobj, element = NULL)
}
\description{
#' @export
lh_post <- function(ytilde, hc, alph, HinvSH_h, t_max, K){
  llike = - 0.5*sum(hc) -
            0.5*t(ytilde) %*% sparseMatrix(i = 1:(t_max*K), j = 1:(t_max*K), x = exp(-hc)) %*% ytilde
  lpost = as.numeric(llike +  -.5*Matrix::t(hc-alph)%*%HinvSH_h%*%(hc-alph))
  return(lpost)
}
}
