% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast.R
\name{recursive_forecast}
\alias{recursive_forecast}
\title{Forecast density of future observations of BVAR model}
\usage{
recursive_forecast(Chain = NULL, y0 = NULL, y_future, t_pred = 12,
  reestimated = F)
}
\arguments{
\item{Chain}{The fatBVARSV object from command BVAR.}

\item{t_pred}{The time prediction horizon.}

\item{y_current}{The current values of y.}

\item{y_obs_future}{The future observable values of y.}
}
\value{
The forecast density of future observations of BVAR model
}
\description{
This function returns a forecast density of future observations of BVAR-SV-fatTail model.
}
\examples{
\dontrun{
forecast_dens1 <- forecast_density(Chain1, y_obs_future)
}

}
