% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simBvarSV.R
\name{sim.VAR.SV}
\alias{sim.VAR.SV}
\title{Simulate data from Gaussian BVAR model with SV}
\usage{
sim.VAR.SV(
  dist,
  K = 5,
  p = 2,
  t_max = 1000,
  b0 = 0.5,
  a0 = 0.5,
  h = 0,
  sigma_h = NULL,
  y0 = matrix(0, ncol = K, nrow = p),
  nu = 6,
  gamma = 0.5,
  sigma_G = NULL,
  seednum = 0,
  burn_in = 0
)
}
\arguments{
\item{dist}{The variable specifies the BVAR error distribution. It should be one of
c("Gaussian","Student","Skew.Student","Skew.Student", "MT","Skew.MT","MST").}

\item{K}{The number of variables in BVAR model.}

\item{p}{The number of lags in BVAR model.}

\item{t_max}{The number of observations}

\item{b0}{The coefficients of B matrix.}

\item{a0}{The coefficients of A matrix.}

\item{h}{The initial log diag(Sigma_t) matrix. sigma_h is set at diag(0.03, K)}

\item{nu}{The degree of freedom.}

\item{gamma}{The skewness from \eqn{[-gamma, gamma]}.}

\item{seednum}{The default seed.}

\item{burn_in}{The discarded observations.}
}
\value{
A list of simulated data with its specification.
}
\description{
This function simulate data from a BVAR model with SV.
\deqn{y_t = B x_t + SQRT(w_t) A^(-1) Sigma_t eps_t}
}
\examples{
\dontrun{
datagen <- sim.VAR.SV(dist="Gaussian")
y <- datagen$y
}
}
