% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterior.R
\name{WAIC}
\alias{WAIC}
\title{WAIC of BVAR model}
\usage{
WAIC(Chain)
}
\arguments{
\item{data}{The fatBVARSV object from command BVAR.}
}
\value{
The WAIC of the fatBVARSVobj at param.
}
\description{
This function returns a WAIC of BVAR-SV-fatTail model.
}
\examples{
\dontrun{
WAIC_chain <- WAIC(Chain)
}
}
