% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterior.R
\name{log_posterior}
\alias{log_posterior}
\title{Posterior distribution of BVAR model}
\usage{
log_posterior(param, data)
}
\arguments{
\item{param}{The one posterior sample of the parameters obtained by fatBVARSVobj$mcmc}

\item{data}{The fatBVARSV object from command BVAR.}
}
\value{
The unnormalized log posterior density of the fatBVARSVobj at param.
}
\description{
This function returns a posterior distribution of BVAR-SV-fatTail model.
}
\examples{
\dontrun{
posterior <- get_posterior(Chain$mcmc[1,], Chain)
}
}
