% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast.R
\name{recursive_forecast}
\alias{recursive_forecast}
\title{Recursive forecast of BVAR model}
\usage{
recursive_forecast(y, t_start = 100, t_pred = 12, K, p,
  dist = "Hyper.multiStudent", SV = T)
}
\arguments{
\item{y}{the whole data matrix T x k.}

\item{t_start}{the start point for forecast.}

\item{t_pred}{The time prediction horizon.}

\item{K}{The number of variables in y.}

\item{p}{The number of lags in BVAR model.}
}
\value{
The recursive forecast density of future observations of BVAR model
}
\description{
This function returns a recursive forecast of future observations of BVAR-SV-fatTail model.
}
\examples{
\dontrun{
recuresive_model1 <- recursive_forecast(y, t_start, t_pred, K, p, dist = "Hyper.multiStudent", SV = T)
}

}
